pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
authors = [{ name = "Hunter Hogan", email = "HunterHogan@pm.me" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Other Audience",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Typing :: Typed",
]
dependencies = [
    "cffconvert",
    "attrs",
    "packaging",
    "requests",
    "ruamel.yaml",
    "tomli",
]
description = ""
keywords = []
license = { 'text' = "CC-BY-NC-4.0" }
name = "updateCitation"
optional-dependencies = { testing = ["pytest-cov", "pytest-xdist", "pytest"] }
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.10"
urls = { Donate = "https://www.patreon.com/integrated", Homepage = "https://github.com/hunterhogan/", Repository = "https://github.com/hunterhogan/" }
version = "0.0.1"

[tool.coverage]
report = { exclude_lines = [
    "if TYPE_CHECKING:",
    "if __name__ == \"__main__\":",
], ignore_errors = false }
run = { branch = true, concurrency = [
    "multiprocessing",
    "thread",
], data_file = "tests/coverage/.coverage", omit = [
    "tests/*",
], parallel = true, source = [
    ".",
] }
xml = { output = "tests/coverage/coverage.xml" }

[tool.pytest.ini_options]
addopts = ["--color=yes", "-n 4"]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
